replace CloudGeneratorTest to CloudDrawerTest
